File hierarchy refactoring
