+++ imports/UI/css/_all.js
@@ -1,0 +1,8 @@
+// Import all CSS files
+import './banners.css';
+import './footer.css';
+import './form.css';
+import './generic.css';
+import './image.css';
+import './navbar.css';
+import './slideshow.css';
