--- imports/UI/loginRequired.js
@@ -1,2 +1,0 @@
-// HTML import
-import './loginRequired.html';
