+++ imports/UI/product/_all.js
@@ -1,0 +1,7 @@
+// Import all the product files
+import './add.js';
+import './edit.js';
+import './banner.js';
+import './page.js';
+
+import './modals/report.js';
