+++ imports/UI/user/_all.js
@@ -1,0 +1,11 @@
+// Import all the user files
+import './contributions.js';
+import './favorite.js';
+import './informations.js';
+import './loginRequired.html';
+import './profile.js';
+
+import './modals/login.js';
+import './modals/register.js';
+import './modals/forgotPassword.js';
+import './modals/resetPassword.js';
